+++ src/uqtestfuns/test_functions/__init__.py
@@ -4,6 +4,7 @@ The init for the 'test_functions' subpackage of UQTestFuns.
 from .ackley import Ackley
 from .borehole import Borehole
 from .bratley1992 import Bratley1992a, Bratley1992b, Bratley1992c, Bratley1992d
+from .cantilever_beam_2d import CantileverBeam2D
 from .damped_oscillator import DampedOscillator, DampedOscillatorReliability
 from .flood import Flood
 from .forrester import Forrester2008
@@ -31,6 +32,7 @@ __all__ = [
     "Bratley1992b",
     "Bratley1992c",
     "Bratley1992d",
+    "CantileverBeam2D",
     "DampedOscillator",
     "DampedOscillatorReliability",
     "Flood",