+++ src/uqtestfuns/test_functions/__init__.py
@@ -5,6 +5,7 @@ from .ackley import Ackley
 from .borehole import Borehole
 from .bratley1992 import Bratley1992a, Bratley1992b, Bratley1992c, Bratley1992d
 from .cantilever_beam_2d import CantileverBeam2D
+from .circular_pipe_crack import CircularPipeCrack
 from .damped_oscillator import DampedOscillator, DampedOscillatorReliability
 from .flood import Flood
 from .forrester import Forrester2008
@@ -35,6 +36,7 @@ __all__ = [
     "Bratley1992c",
     "Bratley1992d",
     "CantileverBeam2D",
+    "CircularPipeCrack",
     "DampedOscillator",
     "DampedOscillatorReliability",
     "Flood",