+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The eight-dimensional damped oscillator reliability problem from
   Der Kiureghian and De Stefano (1990); the problem is based on the existing
   Damped Oscillator model in the code base.
+- The two-dimensional hyper-sphere bound reliability problem
+  from Li et al. (2018).
 - New docs section on list of functions for reliability analysis.
 
 ## [0.2.0] - 2023-06-26