+++ CHANGELOG.md
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Damped Oscillator model in the code base.
 - The two-dimensional hyper-sphere bound reliability problem
   from Li et al. (2018).
+- The two-dimensional cantilever beam reliability problem from
+  Rajashekhar and Ellington (1993).
 - New docs section on list of functions for reliability analysis.
 
 ## [0.2.0] - 2023-06-26