+++ CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   Rajashekhar and Ellington (1993).
 - The two-dimensional four-branch function for reliability analysis
   from Katsuki and Frangopol (1994).
+- The five-dimensional speed reducer shaft reliability problem
+  from Du and Sudjianto (2004).
 - New docs section on list of functions for reliability analysis.
 
 ## [0.2.0] - 2023-06-26